+++ eqcorrscan/tests/lag_calc_test.py
@@ -46,11 +46,13 @@ class TestMethods(unittest.TestCase):
         cls.detections = [Detection(
             detect_time=detection_event.origins[0].time, detect_val=2.0,
             no_chans=5, threshold=1.9, typeofdet='corr', event=detection_event,
-            template_name='test_template'),
+            template_name='test_template', threshold_type='MAD',
+            threshold_input=8.0),
                           Detection(
             detect_time=detection_spicks_event.origins[0].time, detect_val=2.0,
             no_chans=5, threshold=1.9, typeofdet='corr',
-            event=detection_spicks_event, template_name='test_template')]
+            event=detection_spicks_event, template_name='test_template',
+            threshold_type='MAD', threshold_input=8.0)]
         tstart = min(tr.stats.starttime for tr in cls.template)
         cls.delays = [('test_template', [(tr.stats.station, tr.stats.channel,
                                           tr.stats.starttime - tstart)